+++ NEWS.md
@@ -3,6 +3,7 @@
 - `summary()` function now supports a `standardized_coefficients` argument to report either standardized (default) or raw meta-regression coefficients
 - `extract()` function extracts the posterior samples of the model parameters
 - `true_effects()` function to summarize the true effect size estimates
+- `predict()` method for `RoBMA()` and `RoBMA.reg()` models when using the `algorithm = "ss"`
 
 ### Fixes
 - fitting a meta-regression using predictors with missing values result in a clear error message